+++ tests/integration/prepare.js
@@ -9,7 +9,7 @@ const fs = require('fs');
 
 const testsTmpPath = path.resolve(path.join(tempDir, '.__tmp_int'));
 
-const meteorVersion = '1.8.1';
+const meteorVersion = '2.6';
 
 shell.config.fatal = true;
 const appDir = path.join(testsTmpPath, 'test-desktop');
@@ -17,10 +17,11 @@ const appDir = path.join(testsTmpPath, 'test-desktop');
 shell.rm('-rf', testsTmpPath);
 
 if (!fs.existsSync(testsTmpPath) || !fs.existsSync(path.join(appDir, 'package.json'))) {
-    console.log('creating test dir');
+    console.log('creating test dir in ' + testsTmpPath);
     shell.mkdir('-p', testsTmpPath);
-    console.log('creating test meteor app');
+    console.log('creating test meteor app with meteor version=' + meteorVersion);
     shell.exec(`meteor create test-desktop --release=METEOR@${meteorVersion}`, { cwd: testsTmpPath });
+    console.log('created test meteor app');
     const packageJson = JSON.parse(fs.readFileSync(path.join(appDir, 'package.json'), 'utf8'));
     packageJson.dependencies['meteor-desktop'] = path.resolve(path.join(__dirname, '..', '..' ));
     if (process.env.APPVEYOR) {
@@ -32,7 +33,7 @@ if (!fs.existsSync(testsTmpPath) || !fs.existsSync(path.join(appDir, 'package.js
 } else {
     const currentVersion = fs.readFileSync(path.join(appDir, '.meteor', 'release'), 'utf-8').split('@')[1].replace(/[\r\n]/gm, '');
     if (currentVersion !== meteorVersion) {
-        console.log('updating meteor version');
+        console.log('updating meteor version in app dir ' + appDir);
         shell.exec(`meteor update --release=METEOR@${meteorVersion} --all-packages`, { cwd: appDir });
     }
     console.log('meteor npm prune');
@@ -41,9 +42,9 @@ if (!fs.existsSync(testsTmpPath) || !fs.existsSync(path.join(appDir, 'package.js
 console.log('meteor npm install');
 shell.exec('meteor npm install', { cwd: appDir });
 
-
 if (process.env.TRAVIS) {
     shell.config.fatal = false;
+    console.log('adding platform android');
     shell.exec('meteor add-platform android', { cwd: appDir });
     console.log(shell.exec('meteor build ../build --server=127.0.0.1:3000', { cwd: appDir }));
 } else {