+++ plugins/bundler/package.js
@@ -11,7 +11,7 @@ Package.registerBuildPlugin({
     name: 'meteor-desktop-bundler',
     use: ['ecmascript@0.4.4'],
     sources: ['bundler.js'],
-    npmDependencies: { chokidar: '2.0.3' }
+    npmDependencies: { chokidar: '3.5.3' }
 });
 
 Package.onUse(function onUse(api) {