+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 Fixed to work with Meteor 2.6+ Node 14+.
 * Updated `shelljs` to `0.8.5`.
 * Fix bundler plugin to pass non-uglified `code` instead of `undefined` to `fs.writeFileSync` in dev mode.
+- Updated build configs to use Meteor 2.6 when building + testing.
+
+Changes to publish packages under a new name.
+* `meteor-desktop` npm package renamed to `@meteor-community/meteor-desktop`
+* `omega:meteor-desktop-bundler` Atmosphere package renamed to `meteor-community:meteor-desktop-bundler`
+* `omega:meteor-desktop-watcher` Atmosphere package renamed to `meteor-community:meteor-desktop-watcher`
+* Updated all references to these package names in the code.
+* Removed old wojtkowiak build links in README.
+* Removed Contributing and Roadmap README entries.
 
 ## v2.2.5 <sup>24.01.2020</sup>