+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 Fixed to work with Meteor 2.6+ Node 14+.
 * Updated `shelljs` to `0.8.5`.
 * Fix bundler plugin to pass non-uglified `code` instead of `undefined` to `fs.writeFileSync` in dev mode.
-- Updated build configs to use Meteor 2.6 when building + testing.
+* Updated build configs to use Meteor 2.6 when building + testing.
+* Updated `chokidar` dependency from 2.x to 3.x for `meteor-desktop-bundler` and `meteor-desktop-watcher` packages.
 
 Changes to publish packages under a new name.
 * `meteor-desktop` npm package renamed to `@meteor-community/meteor-desktop`